artifacts: allow to set artifacts for different operations in host and topology controller
