+++ pytest_mh/__init__.py
@@ -4,6 +4,7 @@
 
 from __future__ import annotations
 
+from ._private.artifacts import MultihostHostArtifacts, MultihostTopologyControllerArtifacts
 from ._private.data import MultihostItemData
 from ._private.fixtures import MultihostFixture, mh
 from ._private.marks import KnownTopologyBase, KnownTopologyGroupBase, TopologyMark
@@ -21,14 +22,18 @@ from ._private.topology_controller import TopologyController
 
 __all__ = [
     "mh",
+    "KnownTopologyBase",
+    "KnownTopologyGroupBase",
     "MultihostConfig",
     "MultihostDomain",
     "MultihostFixture",
     "MultihostHost",
+    "MultihostHostArtifacts",
     "MultihostHostOSFamily",
     "MultihostItemData",
     "MultihostPlugin",
     "MultihostRole",
+    "MultihostTopologyControllerArtifacts",
     "MultihostUtility",
     "pytest_addoption",
     "pytest_configure",
@@ -36,6 +41,4 @@ __all__ = [
     "TopologyController",
     "TopologyDomain",
     "TopologyMark",
-    "KnownTopologyBase",
-    "KnownTopologyGroupBase",
 ]
